+++ src/main/java/br/com/furafila/mvc/login/service/ILoginService.java
@@ -17,6 +17,6 @@ public interface ILoginService {
 
 	public List<Login> listarEntregador() throws Exception;
 
-	public Long gravarLogin(Login login) throws Exception;
+	public int gravarLogin(Login login) throws Exception;
 
 }
